Fix the build error for suspend debug
../../../../../../kernel/kernel/power/user_sysfs_private.c:182:3:
warning: format '%d' expects argument of type 'int',
but argument 4 has type 'long int' [-Wformat=]
error, forbidden warning: user_sysfs_private.c:182
make[3]: *** [kernel/power/user_sysfs_private.o] Error 1

Reviewed-on: https://gerrit.mot.com/1188825
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key
Tested-by: Jira Key

+++ kernel/power/user_sysfs_private.c
@@ -179,7 +179,7 @@ int tlmm_dump_info(char *buf, int size, int tlmm_num, int tlmm_count)
 			}
 		}
 		pr_debug("%s(), %d\n", __func__, __LINE__);
-		p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
+		p += snprintf(p, size - (p - buf), "(%d)\n", (int)(p - buf));
 	}
 	return p - buf;
 }
@@ -236,7 +236,7 @@ int tlmm_before_sleep_dump_info(char *buf, int size)
 			p += tlmm_dump_cfg(p, size - (p - buf),
 				i, cfg, output_val);
 		}
-		p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
+		p += snprintf(p, size - (p - buf), "(%d)\n", (int)(p - buf));
 	}
 	return p - buf;
 }
@@ -299,7 +299,7 @@ int tlmm_tz_dump_info(char *buf, int size)
 		p += snprintf(p, size - (p - buf), "%d: ", i);
 		p += snprintf(p, size - (p - buf), "%d\n", tz_configs[i]);
 	}
-	p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
+	p += snprintf(p, size - (p - buf), "(%d)\n", (int)(p - buf));
 	return p - buf;
 }
 
@@ -352,7 +352,7 @@ int tlmm_before_sleep_table_dump_info(char *buf, int size)
 		/* cfg &= ~0x40000000; */
 		p += tlmm_dump_cfg(p, size - (p - buf), i, cfg, output_val);
 	}
-	p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
+	p += snprintf(p, size - (p - buf), "(%d)\n", (int)(p - buf));
 	return p - buf;
 }
